fix(test): move fixture-dependent collection test to integration suite
`podkit#test:unit` failed deterministically on a fresh clone. One unit
test — the `collection music` heading annotation case — read generated
audio fixtures via `getStaticFixturesRoot()`, but turbo only declares
`@podkit/test-fixtures#generate-static-fixtures` as a dependency of
`test:integration`; `test:unit` depends on `["build", "^build"]` alone.
With no fixtures on disk the directory scan set exit code 1 and the
assertion on an undefined exit code failed. bunfig's `retry = 2` masked
it as a flake ("attempt 3").

Move the test rather than widening `test:unit`'s dependencies: adding
fixture generation there would make every package's unit tests require
ffmpeg, which is exactly why check-ffmpeg.ts already keeps generation out
of `build`. A sweep of all packages found this to be the only unit test
consuming static fixtures — everything else already lives in
`.integration.test.ts` or e2e.

collection.integration.test.ts needed no new helpers: it already imports
runCollectionMusic, defines an identical runMusic wrapper and
makeContext/makeOut, computes the same AUDIO_FIXTURES_PATH, and calls
ensureFixturesExist() as a guard for running outside the turbo graph.

Also rename the test to say what it asserts. It was called "annotates the
stats heading with playlist name" but twenty lines of commentary
concluded the opposite — the positive case is unreachable without a live
Subsonic server, so it only ever covered the negative case.

Verified by deleting test-packages/test-fixtures/fixtures/ entirely and
running test:unit with --force: 42/42 tasks, 3439 tests, 0 fail.
